requirements.txt
# MCP Server
mcp[cli]==1.23.0
pyyaml==6.0.2

# Document Analysis
pypdf==6.9.1
pdfplumber==0.11.4
python-docx==1.1.2

# Text Analysis
textstat==0.7.4
